Big graphics subsystem refactoring.
* Kernel frame buffer interface now in include.
* Simple color system (ABGR) added.
* Draw functions now take a color.

+++ src/kernel/driver/bcm2835/mailbox/mailbox.c
@@ -39,7 +39,7 @@ void mailbox_write(uint8_t mailbox, void* mail) {
 	
 	}
 	
-	*write = mail;
+	*write = (uint32_t)mail;
 	
 }
